[Silver III] Title: 01타일, Time: 4 ms, Memory: 5016 KB -BaekjoonHub

+++ 백준/Silver/1904. 01타일/01타일.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-long long dp[1000001]={};
+int dp[1000001]={};
 int main(){
     int n;
     scanf("%d",&n);
@@ -11,6 +11,6 @@ int main(){
         dp[i]=(dp[i-1]+dp[i-2])%15746;
 
     }
-
-    printf("%lld",dp[n]);
+    
+    printf("%d",dp[n]);
 }
